Rename CapybaraIntegrationExtensions -> CapybaraSessionExtensions
Fold IntegrationExtensions into SessionExtensions
