--- test/capybara_integration_extensions.rb
@@ -1,33 +1,0 @@
-# frozen_string_literal: true
-
-#
-#  = Integration High-Level Test Helpers
-#
-#  Methods in this class are available to all the integration tests.
-#
-#  login::   Create a session with a given user logged in.
-#  login!::  Same thing,but raise an error if it is unsuccessful.
-#
-#
-module CapybaraIntegrationExtensions
-  # Login the given user in the current session.
-  def login(login = users(:zero_user).login, password = "testpassword",
-            remember_me = true)
-    login = login.login if login.is_a?(User)
-    visit("/account/login")
-
-    fill_in("user_login", with: login)
-    fill_in("user_password", with: password)
-    check("user_remember_me") if remember_me == true
-
-    click_button("Login")
-  end
-
-  # Login the given user, testing to make sure it was successful.
-  def login!(user, *args)
-    login(user, *args)
-    assert_flash(/success/i)
-    user = User.find_by(login: user) if user.is_a?(String)
-    assert_users_equal(user, assigns(:user), "Wrong user ended up logged in!")
-  end
-end
